[usbdev,dv] Fix how we set rand_or_not in subclasses
The previous version of the code would work in C++, where a subclass
initializer overrides the one in the superclass. But it doesn't really
work in SystemVerilog. Instead, the subclass gets a new class variable
called rand_or_not which has a different value from the base class.
Very confusing!

This commit gets rid of the variables from the subclasses and moves
the "init phase customisation" to the start of body(), matching what
some of the vseqs were doing already.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>