[usbdev,dv] Set num_of_bytes earlier in two vseqs
The existing code was pretty unclear. The m_data_pkt.data queue is
created inside call_data_seq() in the base class, where its size is
constrained to num_of_bytes: a variable that is passed in
uninitialised.

I think that the existing implementation was picking a random 7-bit
number to use as an argument, but it was *very* unclear.

This implementation is a little more obvious. I've also constrained
things slightly more to [1, 64], following a review comment from
Mubashir Saleem (thanks).

This change is the same for both usbdev_in_trans_vseq and
usbdev_pkt_sent_vseq.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>