[usbdev,dv] Simplify randomisation in usbdev_av_buffer_vseq
No functional change, but the resulting code is hopefully a little
simpler to read and understand.

Note that this also allows us to get rid of the num_of_bytes class
variable in the subclass. This was shadowing a class variable in the
base class. I think the intention was to constrain the variable in
this instance of the base class to be equal to 8. Since can now pass
it explicitly, there's no need to mess around with these variables at
all.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>