[i2c,rtl] Add TXRST_ON_COND user toggle in Target-Mode
By setting the bit CSR.TARGET_FIFO_CONFIG.TXRST_ON_COND, the block will
then automatically reset the TXFIFO (using TXRST) if it observes either a
STOP/RSTART condition in Target-Mode.

This behaviour may be useful to software, as any remaining data in the TXFIFO
after a Sr/P condition is probably no-longer applicable to the next transfer, so
will likely have to be cleared out anyway. Keeping this behaviour as a toggle
allows software to observe the fifo state before resetting it, which may be
useful to understand how much of the previous transfer completed if that
information is helpful or relevant.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>