[i2c,rtl] NACK next byte on write
When in target mode and receiving an address with a write request, the
ACQ FIFO might not always have space for this. Before this commit the
target would NACK the address, but for SMBus we should ACK the address
and NACK the subsequent byte. That is what this commit implements. The
read requests are unchanged because there are no bytes coming from the
host after that. This does not cause any incompatibility with SMBus
becuase the first address always corresponds to a write.

This commit also adds AcqNackStart as a separate enum value from AcqNack
to communicate the difference between a delayed and an immediate NACK to
the OpenTitan software.

Signed-off-by: Marno van der Maas <mvdmaas+git@lowrisc.org>