[entropy_src] Remove unneeded main SM state and input signals
Now that the CS AES Halt interface is handled by the SHA3 core itself,
the Sha3Quiesce / Sha3MsgDone states can be combined into one state.
This helps reducing the latency of the conditioner and thus the
backpressure onto the entropy pipeline.

The main_stage_rdy_i input signal can be removed as this is identical
to the sha3_state_vld_i input signal checked in Sha3Valid. At this
point in the FSM, it is always asserted and doesn't need to be
checked again.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>