[sram_ctrl] Minor alignment of init and key request behavior
Previously, it was possible to re-request a new key while a memory
init was ongoing. This can result in problematic behavior since
the init done indication would still be set to true afterwards,
which is misleading since part of the array has then been initialized
with the a different key.

This patch changes that and ignores key requests that are issued
while a memory initialization is pending.

On a similar note, it was possible to re-trigger a memory initialization
while one was already ongoing, causing the existing initialization to
restart. This is now also changed so that initialization requests
are ignored while an initialization is still pending.

Signed-off-by: Michael Schaffner <msf@opentitan.org>