[bitstream cache] Make unavailable but required bitstream configurable
By default, the bitstream workspace script creates aliases to vivado
builds for required bitstreams not found in the cache. For CI purposes,
we would like to use a different package. This commit introduces a new
switch so that a different package can be used. For example:

export BITSTREAM="--unavailable=//hw/bitstream/my_alternative_bitstream"

The default is still to use vivado, which is equivalent to

export BITSTREAM="--unavailable=@//hw/bitstream/vivado"

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>