[i2c,rtl] Add extra enum values to ACQDATA reg
This is in preparation for the NACK after hitting a stretch timeout
feature in the I2C target. It also widens the width of
`i2c_acq_byte_id_e` and adds a new value: `AcqNack`.

This commit contains changes based on autogenerated commands. These
are the commands that I ran:
$ ./util/cmdgen.py -u hw/ip/i2c/doc/registers.md
$ ./util/regtool.py -r -t hw/ip/i2c/rtl hw/ip/i2c/data/i2c.hjson

For reviewing you can ignore the autogenerated changes and just look at
`i2c.hjson` and `i2c_pkg.sv`.

Signed-off-by: Marno van der Maas <mvdmaas+git@lowrisc.org>