[pentest] OTBN FI key sideloading & load integrity
This commits adds the following two FI penetration tests for OTBN:
- otbn.fi.load_integrity
- otbn.fi.key_sideload

The first test aims to check whether the CRC checksum over the
DMEM can be manipulated.
The second test aims to test whether a fault can manipulate a key
that gets sideloaded from the key manager.

The host code is located in lowRISC/ot-sca#338

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>