[sival, rv_dm] allow rng seeding in csr_rw and mem_access tests
In CI, generate random seeds using thread_rng and print their values.

A new crate, `rand_chacha` is introduced for this change. Note that this
is what the `rand::prelude::StdRng` uses, but `StdRng` doc mentions that
there're no reproducibility guarantees on `StdRng` and advice to use
`rand_chacha` instead.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>