[bazel] Change default environments for test targets.
By default, `EARLGREY_TEST_ENVS` used to run tests on both test ROM and
regular ROM. This results in a lot of duplicated runs on CI. This commit
removes the test ROM from the defaults list, leaving only the regular
ROM. If a test has explicitly overridden this default value (and e.g.
specified it only runs on test ROM), then it will be unaffected by this
change.

Signed-off-by: Jade Philipoom <jadep@zerorisc.com>