[usbdev,dv] Remove usb_sense_i from usb20_block_if
This signal was the same as usb_vbus.

A reader might reasonably get confused about where the value for
usb_vbus actually comes from(!). It turns out that that it's driven
directly by some code in usb20_driver.sv. Honestly, I think this looks
a bit odd, but that's a problem for a separate patch.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>