[top_earlgrey] Double the ROM size to 64kB
This doubles the ROM size to 64kB.

Note that the base address on the crossbar had to be adapted
since our tooling does not allow for overlapping base / size masks
due to the way the address decoding is implemented in the crossbar.

To that end, the base of the RV_DM ROM is also moved to make room for
larger ROMs at the beginning of the address map.

Signed-off-by: Michael Schaffner <msf@opentitan.org>