[dv,top_level] Fix chip_sw_sysrst_ctrl_ulp_z3_wakeup test
The UVM code in the test sequence should not wait before updating
the phase memory location via backdoor when transitioning to phase
3, or the update can be too late and the test will fail.

More details: In phase 2 the SV side causes sysrst_ctrl to detect
the lid opened, which trigger a wakeup while the CPU is in deep
sleep. SV then calls sync_with_sw, which delays updating the phase
after the CPU enters test_main, and can cause phase 2 to rerun.
Normally this test passes, but the delay can be configured via
sim_dv runtime plusargs and set large enough for this to happen,
and that is what happens for this failure.

The wait in phase 2 is unnecessary since the chip is asleep and
opening the lid causes it to wakeup, so updating the phase can be
done anytime before the CPU restarts running the test code.

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>