[dif] Update hmac dif to enable testing of secure wipe functionality.
Update the hmac finish command to skip disabling of hmac via a flag
argument.

Update the secure wipe call to read the digest registers as most test
cases will require this.

Make the key parameter in mode hmac optional to be able to start a new
hmac operation with the currently loaded key values.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>