[i2c,rtl] Halt controller FSM when unexp. NAK is seen
If the I2C bus NAKs any of our transfers when operating as a
Controller-Transmitter in Byte-Formatted Programming Mode, the current behaviour
is to proceed to the next FDATA entry as normal, disregarding the NAK entirely.
The only side-effect of this event is the block raising the 'nak'
interrupt (though this can be suppressed by the use of NAKOK in the FDATA entry).

This commit changes the behaviour so the Controller-FSM halts upon generation of
the 'nak' interrupt, and proceeds once the interrupt has been acknowledged using
the standard 'Event-Type' interrupt acknowledgement (W1C to INTR_STATE).
The next item in the FMTFIFO is not popped until this acknowledgement is received.
If the NAKOK bit is set for a particular Format Indicator and the bus NAKs our
byte, the 'nak' interrupt is not asserted and the FSM continues to the next
indicator in the fifo.

While the FSM is halted, SCL is released (we halt in the 'ClockPulseAck' state),
and the bus is observed as having an elongated 9th clock period.

Two additional features are also added alongside this change.

- "HOST_NACK_HANDLER_TIMEOUT" csr
This CSR controls a timeout mechanism where if SW does not handle a pending 'nak'
interrupt before the configured timeout elapses, the hardware will automatically
end the active transaction with a STOP condition and disable Host-Mode.

- "STATUS.HOST_DISABLED_NACK_TIMEOUT" bit
This bit is set when the block times out with the "HOST_NACK_HANDLER_TIMEOUT"
mechanism, and clears when SW re-enables Host-Mode. This is intended to inform
software that the timeout occurred.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>