[rv_core_ibex,rtl] Reduce number of performance counters
On top of the RV32I mandatory counters this just leaves the counters for
data memory and instruction memory waits. The others are counting
different types of instructions being executed so are purely a property
of the architectural behaviour of the program so could be found another
way (e.g. via simulation) where the wait counters cannot be easily
modelled.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>