[prim] Weaken "known" assertion in prim_fifo_sync
The existing assertion was false because it essentially asserted that
we never write an unknown value into the FIFO. Not something that can
be asserted from within the FIFO (obviously) and not actually true in
the design, because we use the FIFO on the output from memories.

This assertion is slightly weaker. Instead of saying "rdata is never
X", it tracks whether we've seen an X as input since the last time the
FIFO was empty. If we haven't, it asserts that rdata is not X.

Running locally, Jasper doesn't quickly find a proof for this new
statement (so I'm absolutely sure there's a better way to do this).
But it also doesn't find a counterexamples, so I'm convinced this is
an improvement!

I'm not sure whether the "ifdef FPV_ON" is needed here. It makes it
more obvious that this definitely won't appear in the synthesized
design, but I don't have strong feelings: maybe it's something that
makes sense to track in simulations too.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>