[csrng] Perform debug state reads directly from state database
Previously, an additional flopped copy of the state of the selected
instance was used for this purpose. Reading directly from the state
database allows saving 384 FFs. The timing should not increase
significantly due to this change. The only logic between the state
database and the INT_STATE_VAL register readable by software is the
reduction tree required to get 32 bits out of the 3x(384 + 2) state
bits.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>