[bazel] Add constant definitions for sigverify.
Keys stored in OTP will have to be configured with constant values that
are defined in the rom sigverify library. This change adds such constant
definitions to //rules/const.bzl.

In addition to this. The SPX+ config id enums were updated to use higher
hamming distance values and to use a name reflecting the SPX+ parameter
configuration.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>