[dv,usbdev] Fix usbev_pkt_received and usbdev_enable
The 'pkt_received' interrupt is now of type Status so it is
asserted only whilst the RX FIFO still contains the packet
description.
Check the properties and content of the received packet in
each sequence.
Commute the 4n constraint on the length of the received packets,
pending more thorough checking.
Keep the 'recover_orig_data' function local to the affected
sequences until such time as we can eliminate/reduce the
reordering being performed.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>