[dv,usbdev] Data toggle restore sequence
Add a test sequence for the additional Data Toggle restore
functionality.
This sequence first performs some basic tests using the CSR
interface and then brings up the USB connection and tests their
interaction with the USB traffic.

Note that the data toggles may safely be changed between packets
in the DV environment because the USB traffic is controlled.
For active endpoints on a live system they must be modified only
under very specific conditions such as halt or reset.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>