[ottf] Add support for CRC protected host->dev ujson messages
We current only protect device to host messages with CRC but not
the other way around. This PR introduces a mecanism for the other
direction. On the device, it suffices to replace a call like
  ujson_deserialize_my_struct(ujctx, data)
by
  UJSON_WITH_CRC(ujson_deserialize_my_struct, ujctx, data)

Behind the scenes, this will receive the message and then receive
another ottf-specific ujson message containing the crc:
  {crc: 123345}
It then checks that the CRC of the whole first message matches.

On the host send, introduce a new `send_with_crc` method in the
`UartSend` trait that sends the message and then the CRC.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>