[ottf] Enable CRC for ujson memory backdoor messages
With this PR, the ottf framework now expects the `mem` ujson
messages to be CRC protected. The `OTTF_WAIT_FOR` macro is updated
to do that so it is automatic for users of this macro. On the host
side, the test utils are updated to match the new behaviour.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit a5998e0d8e241122afbfe0cb0f683e391ca7202d)