[csrng/rtl] Add command out of sequence cmd_sts error
This commit adds a command out of sequence error to the cmd_sts signal
going to the consumers as a command response. The cmd_sts signal will
now signal an error when a command is issued out of sequence. This can
mean one of two things. Either an instantiate command is issued on an
already instantiated state or any command other than an instantiate is
issued before the state was instantiated.

This commit also adds the documentation and DV alignment for
the rtl changes.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>