[otbn, rtl] Fix X-propagation issue in prefetch stage
When executing near Xs in imem it's possible for Xs to be received on
the incoming imem data. The prefetcher looks at the incoming data to see
if its a branch instruction.

Previous to this change this branch check took priority over a check
related to looping which altered the request going to imem. So when the
incoming imem data was X, when simulation propagated Xs though ifs we
end up with an X on the imem request address.

With this change the check for the branch happens regardless of whether
or not the looping related logic activates and the branch logic itself
does not alter the imem request so it avoids generating Xs on the
external (to OTBN core) imem request interface.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>