[cip,dv] Fix nitty style problem in cip_base_vseq
No functional change, but VCS complains when compiling the code that
was there before, saying that "The SystemVerilog LRM requires a
non-empty 'for_initialization'."

I'm not completely convinced that VCS is right! Looking the "Loop
statements" section of IEEE 1800, the syntax given is:

    for ( [ for_initialization ] ; [ expression ] ; [ for_step ] )

Looking at the "Syntactic description" section near the top of the
document, I see that stuff in square brackets denotes an optional
item.

But it's nicer not to get warnings from the build step, so it's
probably worth matching what VCS expects :-)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>