[spi_host,dv] Avoid dubious typing in spi_host_scoreboard.sv
The existing code was assigning the result of getc() (which is a byte)
to a string and then calling atoi() on the result. It turns out that
VCS and Xcelium both do something sensible in this situation, but at
least VCS gives a warning at build time.

Do the (much) simpler trick of implementing atoi by subtracting '0'
:-)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>