[flash_ctrl] Revert back to using per-bank, dedicated scrambling modules
An inspection of the DV issues arising from the change to share a single
scrambling module between Flash banks revealed that there are indeed
some corner cases that are not properly handled by the current design.
Fixing those corner cases is tricky and may lead to further issues.

For this reason, this commit reverts the following commits for the time
being:

- [doc,flash_ctrl] Fix inconsistency and typo in documentation
  3a35e26c7a1e6e26aa38f4ae899670274f69b1d1

- [flash_ctrl] Use redundant arbiters in scrambling module
  a63c2d9a7abeef6411b9746f7540f444217aefa4

- [flash_ctrl] Share flash_scrambling module among banks
  c95dc32e6235cd0dd3beaee313dc54e25cebd8b4

This is related to lowRISC/OpenTitan#22089.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>