[usbdev,dv] Move covergroup descriptions to testplan
This is a pretty mechanical change: it is basically copying the list
of covergroups from usbdev_env_cov.sv into usbdev_testplan.hjson. I've
taken the long comment above each covergroup and turned it into the
description of the covergroup in the testplan.

The only nontrivial change is to a covergroup name (pid_type_end_cg),
which had a stray "f".

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>