[ot_certs] Refactor extension handling
With this commit, the certificate no longer inlines the DICE TCB
field but instead has support for arbitrary extension. At the moment
only the DICE TCB extension is implemented but it will be much easier
to add more in the future. This commit also refactors the DICE TCB
info extension by removing the DiceTcbInfo in dice_tcb.rs in favor
of using the DiceTcbInfoExtension in the template. As a result,
the builder is no longer capable of outputing some DICE TCB
fields (index, vendorInfo and type) but those are unused at the moment
and can be brought back if needed.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>