[hmac] Coding style fixes
This fixes existing coding style issues from earlier in the HMAC RTL
to bring it the current coding style guide. It also adds minor fixes
for issues raised in the reviews for the wider digest PR, and fixes
the I/O port suffixes in hmac_core.sv per the coding style.

Signed-off-by: Ghada Dessouky <gdessouky@google.com>