[dvsim] Be more explicit about seeds when running tests
This should avoid ugly paths all over the place and tests that are
hard to reproduce. We still get the long build seeds that were the
motivation for the 256 bit constants that we had before.

The base seeds are now all 16 bit values, so should be a bit more
convenient to type on the command line. But the build seeds should be
much longer (256 bits) if needed.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>