[usbdev,env] DV env changes to avoid csr_test failures
This draft pull request introduces changes to the ubsdev environment to facilitate the CSR tests failures, specifically the csr_bit_bash and mem_walk.

Changes include:

Addition of the driver_enable task in usb20_block_if to manage the driver by enabling and disabling it.
The driver is now set to a disabled state by default, ensuring it remains inactive when not required, such as during CSR tests.
To activate the driver when necessary, the driver_enable function can be called in a specific sequence, within the pre_task method, enabling the driver for USB transactions.

Signed-off-by: Mubashir <mubashir.impare@gmail.com>