[spdx] Extract license for OTBN dcrypto library
Moving this license to the `LICENSES/` directory allows it to be
discovered by automated tools.

The `LicenseRef-...` SPDX headers are the canonical way of referring to
custom license files. These are required because `dcrypto` uses a
modified version of the BSD-3-Clause license that explicitly uses Google
instead of being generic.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>