[cert] add cert lib with util functions / structs
This adds a cert library in the silicon_creator dir that contains helper
functions and structs for managing OpenTitan certificates. Specifically
a function is added to check if a cert's serial number matches what is
expected. This function will be used to determine whether CDI0/1 certs
need to be updated in the ROM_EXT (if the ROM_EXT or OWNER firmware is
ever updated).

Signed-off-by: Tim Trippel <ttrippel@google.com>