[ci] Add reuse lint check to CI
Replaces the existing license checker with the FSFE's REUSE tool.

The `--suppress-deprecation` flag is required because the tool authors
have not yet decided on whether the `.reuse/dep5` file or comments
within files should take precedence. Many many warnings are emitted
about this not being supported in the future, but there is no workaround
at the moment so we have to suppress the warnings.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>