[ot_certs] Add support for certificate validity dates
Thsi commit enables the template to sepcify the not_before and
not_after date/time. The code assumes that the format is the ASN1
GeneralizedTime. This causes a slight issue which is that the
validity can also be representd as a UtcTime so the parser might
be bail out certificates produced externally and that use UtcTime
(openssl will produce such certificate). This issue could be solve
by converting UtcTime to GeneralizedTime when parsing but the openssl
rust binding does not export the required function. Since this is a
very minor issue, leave that as a future fix.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit 263be3935b6715ee85a747ce4a500f4c8f069321)