[keymgr_dpe, dv] fixes issue where exp_kmac_key for sideload keys
When a sideload key for kmac was never generated in tests due to randomization
there was no mechanism to reset the exp_kmac_key to 0 as it should be. So in some
seeds after disabling keymgr_dpe and checking sideload keys a comparison was being done
against the exp_kmac key and the kmac_key in rtl which was a 0. The exp_kmac key would
not be properly set to 0 and a false fail was happening. This commit resets the exp_kmac
key to 0 when there is no sideload key generated.

Signed-off-by: Bobby Reyes <bobby@rivosinc.com>