[doc,usbdev] Separated Available Buffer FIFOs
In place of the single Available Buffer FIFO there are
now two separate FIFOs holding buffer IDs for use by SETUP
packets and by OUT packets.
This allows software to ensure that there is always a
buffer available exclusively for the reception of SETUP
packets which are of much greater priority on the USB,
even under heavy OUT traffic.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>