[usbdev] Fixup test name with 'usbdev_' prefix
This caused breakage with some of our CI metrics gathering scripts, as by
convention all tests start with the blockname.

This shouldn't be an issue, but it turns out there is strange underlying dvsim
behaviour where if a testpoint contains no tests, it creates a 'dummy' test
output with 0 total runs, but with the same name as the testpoint. This dummy
name clashes the name of the unmapped test in this case.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>