[i2c] Clean up automatic Stop generation and NACK handling
Consolidate NACK handling events into a single "controller halt" status
interrupt. The event bits in the CONTROLLER_EVENTS CSR combine to drive
the interrupt. Clear the bits to acknowledge the interrupt.

Change up the NACK handling to finish the current byte, force transition
to the Idle state, and hold there, as well as beginning the timeout
counter's incrementing there. SCL will be held low if the transaction
was not stopped already.

Track automatic Stop generation to eliminate races with software, since
changes to the control bits could occur while the Stop symbol was being
generated, leading to a transition to the PopFmtFifo state and the
discarding of an unused FMT FIFO entry. Proceed directly to Idle if the
current Stop symbol was automatically generated.

Signed-off-by: Alexander Williams <awill@opentitan.org>